backend/requirements.txt
fastapi[standard]
sqlmodel
openai==1.3.0  
python-dotenv